refactor: enum 객체들을 constant 패키지로 이동시켜라
