+++ src/main/java/domain/AttendanceSystem.java
@@ -1,13 +1,12 @@
 package domain;
 
 import domain.constants.ErrorMessage;
+import domain.constants.ExpulsionStatus;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
