+++ src/main/java/domain/constants/AnswerCommand.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.constants;
 
 import java.util.Arrays;
 import java.util.Objects;
