+++ src/main/java/domain/constants/ExpulsionStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.constants;
 
 import java.util.Arrays;
 import java.util.Comparator;
